Move place_name function into validate_data module.
