+++ validate_data.py
@@ -1,5 +1,7 @@
 """A module to validate fetched data values."""
+import json
 from flask import render_template
+from luts import json_types
 
 from fetch_data import (
     add_titles
@@ -163,3 +165,35 @@ def get_poly_3338_bbox(gdf, poly_id):
 def is_di_empty(di):
     if len(di) == 0:
         return 404  # http status code
+
+def place_name(place_type, place_id):
+    """
+    Determine if provided place_id corresponds to a known place.
+
+    Args:
+        place_type (str): point, huc, or pa
+        place_id (str): place identifier (e.g., AK124)
+
+    Returns:
+        Name of the place if it exists, otherwise None
+    """
+    if place_type == "point":
+        f = open(json_types["communities"], "r")
+    elif place_type == "huc":
+        f = open(json_types["huc8s"], "r")
+    elif place_type == "pa":
+        f = open(json_types["protected_areas"], "r")
+    else:
+        return None
+
+    places = json.load(f)
+    f.close()
+
+    for place in places:
+        if place_id == place["id"]:
+            full_place = place["name"]
+            if "alt_name" in place and place["alt_name"] is not None:
+                full_place += " (" + place["alt_name"] + ")"
+            return full_place
+
+    return None
